+++ driver/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.1.2
+- Supporting dynamic window resizing for Windows
+
 ## 0.1.1+2
 - Enabled iOS fastdriver tests