+++ tool/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.3.0+3
+-  Supporting dynamic window resizing for Windows
+
 ## 0.3.0+2
 - Supporting changing resolution on Windows in Flutter 1.19.+